Adds a location label, and starts loading stuff
Adds a label which tell the user their current location (currently hard
coded at Urbana, IL). Also adds the start of loading displays. So when
the app starts determining location, it will tell the user. When it
starts loading images, it will tell the user. This doesnt actually
happen currently, but I have the framework in place.